pyproject.toml
[project]
name = "pydantic-checks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "google-generativeai>=0.8.5",
    "pydantic>=2.12.4",
]
